Added cpp files to the hh files
